[Fix] 분리했던 파일 병합
- 문제의 원인이 같은 이름의 파일을 각자다른 두 브랜치에서 만든 것이었습니다.
- 파일의 해시값을 알기가 어려워서 일단 로컬에서 파일명을 바꾼뒤에 머지하면서 충돌이 나는 동일한 이름의 파일들을 정리하고, 바꾼이름의 파일에서 필요한 내용들을 머지된 파일에 옮겨주었습니다.
