+++ MateRunner/MateRunner/Domain/UseCase/Home/DefaultRunningUseCase.swift
@@ -11,10 +11,18 @@ import RxSwift
 
 final class DefaultRunningUseCase: RunningUseCase {
     private let coreMotionManager = CoreMotionManager()
-    private let disposeBag = DisposeBag()
+	var runningTimeSpent: BehaviorSubject<Int> = BehaviorSubject(value: 0)
+	var cancelTimeLeft: BehaviorSubject<Int> = BehaviorSubject(value: 3)
+	var popUpTimeLeft: BehaviorSubject<Int> = BehaviorSubject(value: 2)
+	var inCancelled: BehaviorSubject<Bool> = BehaviorSubject(value: false)
+	var shouldShowPopUp: BehaviorSubject<Bool> = BehaviorSubject(value: false)
     var distance = BehaviorSubject(value: 0.0)
     var progress = BehaviorSubject(value: 0.0)
     var finishRunning = BehaviorSubject(value: false)
+	private var runningTimeDisposeBag = DisposeBag()
+	private var cancelTimeDisposeBag = DisposeBag()
+	private var popUpTimeDisposeBag = DisposeBag()
+	private let disposeBag = DisposeBag()
 
     func executePedometer() {
         self.coreMotionManager.startPedometer()
@@ -26,6 +34,42 @@ final class DefaultRunningUseCase: RunningUseCase {
             })
             .disposed(by: self.disposeBag)
     }
+	
+	func executeTimer() {
+		self.generateTimer()
+			.bind(to: self.runningTimeSpent)
+			.disposed(by: self.runningTimeDisposeBag)
+	}
+	
+	func executeCancelTimer() {
+		self.generateTimer()
+			.subscribe(onNext: { [weak self] newTime in
+				self?.shouldShowPopUp.onNext(true)
+				self?.checkTimeOver(from: newTime, with: 3, emitTarget: self?.cancelTimeLeft) {
+					self?.inCancelled.onNext(true)
+					self?.cancelTimeDisposeBag = DisposeBag()
+				}
+			})
+			.disposed(by: self.cancelTimeDisposeBag)
+	}
+	
+	func executePopUpTimer() {
+		self.generateTimer()
+			.subscribe(onNext: { [weak self] newTime in
+				self?.shouldShowPopUp.onNext(true)
+				self?.checkTimeOver(from: newTime, with: 2, emitTarget: self?.popUpTimeLeft) {
+					self?.shouldShowPopUp.onNext(false)
+					self?.popUpTimeDisposeBag = DisposeBag()
+				}
+			})
+			.disposed(by: self.popUpTimeDisposeBag)
+	}
+	
+	func invalidateCancelTimer() {
+		self.cancelTimeDisposeBag = DisposeBag()
+		self.shouldShowPopUp.onNext(false)
+		self.cancelTimeLeft.onNext(3)
+	}
 }
 
 // MARK: - Private Functions
@@ -47,4 +91,26 @@ private extension DefaultRunningUseCase {
         // *Fix : 0.05 고정 값 데이터 받으면 변경해야함
         self.progress.onNext(self.convertToKilometer(value: value) / 0.05)
     }
+	
+	func checkTimeOver(
+		from time: Int,
+		with limitTime: Int,
+		emitTarget: BehaviorSubject<Int>?,
+		actionAtLimit: () -> Void
+	) {
+		guard let emitTarget = emitTarget else { return }
+		emitTarget.onNext(limitTime - time)
+		if time >= limitTime {
+			actionAtLimit()
+		}
+	}
+	
+	func generateTimer() -> Observable<Int> {
+		return Observable<Int>
+			.interval(
+				RxTimeInterval.seconds(1),
+				scheduler: MainScheduler.instance
+			)
+			.map { $0 + 1 }
+	}
 }
