+++ MateRunner/MateRunner/Presentation/HomeScene/ViewModel/SingleRunningViewModel.swift
@@ -8,16 +8,24 @@
 import Foundation
 
 import RxSwift
+import RxRelay
 
 final class SingleRunningViewModel {
     struct Input {
-        let viewDidLoadEvent: Observable<Void>
+		let viewDidLoadEvent: Observable<Void>
+		let finishButtonLongPressDidBeginEvent: Observable<Void>
+		let finishButtonLongPressDidCancelEvent: Observable<Void>
+		let finishButtonDidTapEvent: Observable<Void>
     }
     struct Output {
         @BehaviorRelayProperty var distance: Double?
         @BehaviorRelayProperty var progress: Double?
         @BehaviorRelayProperty var calorie: Int?
         @BehaviorRelayProperty var finishRunning: Bool?
+		@BehaviorRelayProperty var timeSpent: String = ""
+		@BehaviorRelayProperty var navigateToResult: Bool = false
+		var cancelTime: PublishRelay<String> = PublishRelay<String>()
+		var isToasterNeeded: PublishRelay<Bool> = PublishRelay<Bool>()
     }
 
     let runningUseCase: DefaultRunningUseCase
@@ -34,7 +42,48 @@ final class SingleRunningViewModel {
                 self?.runningUseCase.executePedometer()
             })
             .disposed(by: disposeBag)
-        
+		input.viewDidLoadEvent
+			.subscribe(onNext: { [weak self] in
+				self?.runningUseCase.executeTimer()
+			})
+			.disposed(by: disposeBag)
+		
+		input.finishButtonLongPressDidBeginEvent
+			.subscribe(onNext: { [weak self] in
+				self?.runningUseCase.executeCancelTimer()
+			})
+			.disposed(by: disposeBag)
+		
+		input.finishButtonLongPressDidCancelEvent
+			.subscribe(onNext: { [weak self] in
+				self?.runningUseCase.invalidateCancelTimer()
+			})
+			.disposed(by: disposeBag)
+		
+		input.finishButtonDidTapEvent
+			.subscribe(onNext: { [weak self] in
+				self?.runningUseCase.executePopUpTimer()
+			})
+			.disposed(by: disposeBag)
+		
+		self.runningUseCase.cancelTimeLeft
+			.map({ $0 == 3 ? "종료" : "\($0)" })
+			.bind(to: output.cancelTime)
+			.disposed(by: disposeBag)
+		
+		self.runningUseCase.inCancelled
+			.bind(to: output.$navigateToResult)
+			.disposed(by: disposeBag)
+		
+		self.runningUseCase.runningTimeSpent
+			.map(self.convertToTimeFormat)
+			.bind(to: output.$timeSpent)
+			.disposed(by: disposeBag)
+		
+		self.runningUseCase.shouldShowPopUp
+			.bind(to: output.isToasterNeeded)
+			.disposed(by: disposeBag)
+		
         self.runningUseCase.distance
             .bind(to: output.$distance)
             .disposed(by: disposeBag)
@@ -49,4 +98,16 @@ final class SingleRunningViewModel {
         
         return output
     }
+	
+	private func convertToTimeFormat(from seconds: Int) -> String {
+		func padZeros(to text: String) -> String {
+			if text.count < 2 { return "0" + text }
+			return text
+		}
+		let hrs = padZeros(to: String(seconds / 3600))
+		let mins = padZeros(to: String(seconds % 3600 / 60))
+		let sec = padZeros(to: String(seconds % 3600 % 60))
+		
+		return "\(hrs):\(mins):\(sec)"
+	}
 }
